+++ include/nanobind/stl/unordered_set.h
@@ -8,8 +8,7 @@ NAMESPACE_BEGIN(detail)
 
 template <typename Key, typename Hash, typename Compare, typename Alloc>
 struct type_caster<std::unordered_set<Key, Hash, Compare, Alloc>>
-    : set_caster<std::unordered_set<Key, Hash, Compare, Alloc>, Key>
-{
+    : set_caster<std::unordered_set<Key, Hash, Compare, Alloc>, Key> {
 };
 
 NAMESPACE_END(detail)