+++ tests/test_stl.cpp
@@ -8,8 +8,9 @@
 #include <nanobind/stl/optional.h>
 #include <nanobind/stl/variant.h>
 #include <nanobind/stl/map.h>
-#include <nanobind/stl/unordered_set.h>
 #include <nanobind/stl/array.h>
+#include <nanobind/stl/unordered_set.h>
+#include <nanobind/stl/set.h>
 
 NB_MAKE_OPAQUE(NB_TYPE(std::vector<float, std::allocator<float>>))
 
@@ -304,13 +305,20 @@ NB_MODULE(test_stl_ext, m) {
 
     // ----- test58-test62 ------ */
     m.def("set_return_value", []() {
+        std::set<std::string> x;
+        for (int i = 0; i < 10; ++i)
+            x.emplace(std::string(1, 'a' + i));
+        return x;
+    });
+
+    m.def("unordered_set_return_value", []() {
         std::unordered_set<std::string> x;
         for (int i = 0; i < 10; ++i)
             x.emplace(std::string(1, 'a' + i));
         return x;
     });
-    m.def(
-        "set_in_value", [](std::unordered_set<std::string> x) {
+
+    m.def("set_in_value", [](std::set<std::string> x) {
             if (x.size() != 10)
                 fail();
             for (int i = 0; i < 10; ++i) {
@@ -320,8 +328,10 @@ NB_MODULE(test_stl_ext, m) {
             }
         },
         nb::arg("x"));
+
+
     m.def(
-        "set_in_lvalue_ref", [](std::unordered_set<std::string>& x) {
+        "unordered_set_in_value", [](std::unordered_set<std::string> x) {
             if (x.size() != 10)
                 fail();
             for (int i = 0; i < 10; ++i) {
@@ -331,8 +341,21 @@ NB_MODULE(test_stl_ext, m) {
             }
         },
         nb::arg("x"));
+
     m.def(
-        "set_in_rvalue_ref", [](std::unordered_set<std::string>&& x) {
+        "set_in_lvalue_ref", [](std::set<std::string>& x) {
+            if (x.size() != 10)
+                fail();
+            for (int i = 0; i < 10; ++i) {
+                std::string key(1, 'a' + i);
+                if (x.find(key) == x.end())
+                    fail();
+            }
+        },
+        nb::arg("x"));
+
+    m.def(
+        "set_in_rvalue_ref", [](std::set<std::string>&& x) {
             if (x.size() != 10)
                 fail();
             for (int i = 0; i < 10; ++i) {